Nuevo splash, nuevp efecto, Servicio de Notificaciones con eliminacion, terminados.

+++ app/src/main/java/intellisoft/bo/com/intellibusiness/Entity/Principal/Entity.java
@@ -9,16 +9,16 @@ import intellisoft.bo.com.intellibusiness.entity.anotation.Key;
 public class Entity implements Serializable, Cloneable {
 
 	@Key
-	protected Long id;
+	protected int id;
 
 	@Ignore
 	private Action action = Action.None;
 
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
